Node-JS

+++ projects/Pr 8. Make API of Project/controllers/BlogController.js
@@ -31,6 +31,12 @@ const addBlog = async (req, res) => {
 const adminViewBlog = async (req, res) => {
   try {
     const allBlogs = await blogModel.find({}).populate('authorId');
+    if (!allBlogs.length) {
+      return res.status(200).send({
+        success: true,
+        message: "Blogs Not Available"
+      });
+    }
     return res.status(200).send({
       success: true,
       message: "All Blogs",
@@ -47,6 +53,12 @@ const adminViewBlog = async (req, res) => {
 const userViewBlog = async (req, res) => {
   try {
     const blogsOfUser = await blogModel.find({ authorId: req.user?._id });
+    if (!blogsOfUser.length) {
+      return res.status(200).send({
+        success: true,
+        message: "Blogs Not Available"
+      });
+    }
     return res.status(200).send({
       success: true,
       message: "Blogs Of User",
@@ -83,6 +95,13 @@ const userDeleteBlog = async (req, res) => {
   try {
     let id = req.query.id;
     let single = await blogModel.findById(id);
+    let checkAuthor = await blogModel.find({ authorId: single?.authorId });
+    if (!checkAuthor.length) {
+      return res.status(400).send({
+        success: false,
+        message: "Blog Not Found"
+      });
+    }
     if (fs.existsSync(single?.image)) {
       fs.unlinkSync(single?.image);
     }
@@ -98,7 +117,77 @@ const userDeleteBlog = async (req, res) => {
     });
   }
 }
+const adminUpdateBlog = async (req, res) => {
+  try {
+    let { id, title, description } = req.body;
+    let single = await blogModel.findById(id);
+    if (req.file) {
+      if (fs.existsSync(single?.image)) {
+        fs.unlinkSync(single?.image);
+      }
+      await blogModel.findByIdAndUpdate(id, {
+        title: title,
+        description: description,
+        image: req.file?.path
+      });
+    } else {
+      await blogModel.findByIdAndUpdate(id, {
+        title: title,
+        description: description,
+        image: single?.image
+      });
+    }
+    return res.status(200).send({
+      success: true,
+      message: "Blog Updated"
+    });
+  } catch (err) {
+    return res.status(500).send({
+      success: false,
+      message: err
+    });
+  }
+}
+const userUpdateBlog = async (req, res) => {
+  try {
+    let { id, title, description } = req.body;
+    let single = await blogModel.findById(id);
+    let checkAuthor = await blogModel.find({ authorId: single?.authorId });
+    if (!checkAuthor.length) {
+      return res.status(400).send({
+        success: false,
+        message: "Blog Not Found"
+      });
+    }
+    if (req.file) {
+      if (fs.existsSync(single?.image)) {
+        fs.unlinkSync(single?.image);
+      }
+      await blogModel.findByIdAndUpdate(id, {
+        title: title,
+        description: description,
+        image: req.file?.path
+      });
+    } else {
+      await blogModel.findByIdAndUpdate(id, {
+        title: title,
+        description: description,
+        image: single?.image
+      });
+    }
+    return res.status(200).send({
+      success: true,
+      message: "Blog Updated",
+      checkAuthor
+    });
+  } catch (err) {
+    return res.status(500).send({
+      success: false,
+      message: err
+    });
+  }
+}
 
 module.exports = {
-  addBlog, adminViewBlog, userViewBlog, adminDeleteBlog,userDeleteBlog
+  addBlog, adminViewBlog, userViewBlog, adminDeleteBlog, userDeleteBlog, adminUpdateBlog, userUpdateBlog
 }
